Fixed Data Fetching Issue

+++ backend/app/models/product.py
@@ -1,6 +1,6 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, ForeignKey
 from app.db.database import Base
-from sqlalchemy import ForeignKey
+
 
 class Product(Base):
     __tablename__ = "products"
@@ -8,7 +8,6 @@ class Product(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String)
     description = Column(String)
-    price = Column(Float)
+    price = Column(Integer)
     quantity = Column(Integer)
     user_id = Column(Integer, ForeignKey("users.id"))
-
